Merge branch 'NPBT-NG' of https://github.com/NMarkgraf/NPBT into NPBT-NG

# Conflicts:
#	ReleaseNotes.md
#	src/beamercolorthemeNPBT-EUFOM.sty
#	src/beamercolorthemeNPBT-FOM-ifes.sty
#	src/beamercolorthemeNPBT-FOM.sty
#	src/beamercolorthemeNPBT-SC.sty
#	src/beamerouterthemeNPBT-FOM-ifes.sty
#	src/beamerouterthemeNPBT-FOM.sty
#	src/beamerthemeNPBT.sty
#	src/example/NPBT_exsample.tex
#	src/images/NPBT-FOM-ifes-background.png
#	src/images/NPBT-FOM-ifes-frametitlebackrgound.png
#	src/images/NPBT-SC-linie.png
#	src/images/NPBT-eufom-background.png
#	src/images/NPBT-eufom-frametitlebackground.png
#	src/images/NPBT_FOM_ifes_backgound.png
#	src/images/NPBT_FOM_ifes_background.png
#	src/images/NPBT_FOM_ifes_frametitlebackgound.png
#	src/images/NPBT_FOM_ifes_frametitlebackground.png
#	src/images/NPBT_SC_linie.png
#	src/images/NPBT_eufom_backgound.png
#	src/images/NPBT_eufom_background.png
#	src/images/NPBT_eufom_frametitlebackgound.png
#	src/images/NPBT_eufom_frametitlebackground.png
#	src/images/lNPBT_SC_linie.png

+++ makefile
@@ -2,29 +2,53 @@
 # Makefile (C) in 2018 by Norman Markgraf
 # ========
 # Version 2.0 	nm (15. Aug. 2018) 	New first release
-MAKEFILEVERSION=2.0
+# Version 2.1 	nm (10. Okt. 2018) 	some updates
+MAKEFILEVERSION=2.1
 #
 # 
 #
 # ######################################################################################
-
-
+RELEASE="v4.1.2"
+RELDATE="2018/10/10"
+# ######################################################################################
 PKGNAME=beamertheme-npbt
 SRC=src
 BUILD=build
-PKGBUILD=$(BUILD)/$(PKGNAME)
+EXAMPLE=examples
+PKGBUILD=$(BUILD)/$(PKGNAME)-$(RELEASE)
+EXBUILD=$(BUILD)/$(EXAMPLE)
+RSCRIPT=Rscript
+
+build-examples: build
+	mkdir $(EXBUILD)
+	cp -R $(PKGBUILD)/* $(EXBUILD)
+	cp -R $(EXAMPLE)/* $(EXBUILD)
+	cd $(EXBUILD); $(RSCRIPT) ./render_tests.R
 
 build:
-	mkdir $(PKGBUILD)
-	cp -R $(SRC)/* $(PKGBUILD)
-	zip 
+	mkdir --mode=a=wrx $(PKGBUILD)
+	cp -R --no-preserve=all $(SRC)/* $(PKGBUILD)
+	cp ReleaseNotes.md $(PKGBUILD)
+	cp README.md $(PKGBUILD)
+	cd $(PKGBUILD); pandoc -i README.md -o README.html
+	cd $(PKGBUILD); pandoc -i ReleaseNotes.md -o ReleaseNotes.html
+	cd $(PKGBUILD); sed -i -e "s_YYYY/MM/DD vX.X.X_$(RELDATE) $(RELEASE)_g" *.sty;
+	if [ -f $(PKGBUILD)/*.sty-e ]; then cd $(PKGBUILD); rm *.sty-e; fi
+	cd $(PKGBUILD); chmod -R a+rwx *
+	cd $(BUILD); zip -r $(PKGNAME)-$(RELEASE).zip $(PKGNAME)-$(RELEASE)
 
-clean:
-	rm -r $(PKGBUILD)
+clean-build:
+	if [ -d $(PKGBUILD) ]; then rm -r $(PKGBUILD); fi
+
+clean-examples:
+	if [ -d $(EXBUILD) ]; then rm -r $(EXBUILD); fi
+
+clean: clean-build clean-examples
+	@echo "Done."
 
 # --------------------------------------------------------------------------------------
 #
-.PHONY: build clean
+.PHONY: build build-examples clean
 
 # --------------------------------------------------------------------------------------
 # EoF
